refactor(api): move route in team bounded context
DELETE /api/certification-center-invitations/{certificationCenterInvitationId}
