+++ api/lib/application/certification-center-invitations/index.js
@@ -2,7 +2,6 @@ import Joi from 'joi';
 
 import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
-import { certificationCenterInvitationAdminController } from '../../../src/team/application/certification-center-invitation/certification-center-invitation.admin.controller.js';
 import { certificationCenterInvitationController } from './certification-center-invitation-controller.js';
 
 const register = async function (server) {
@@ -49,29 +48,6 @@ const register = async function (server) {
           "- Cette route permet de récupérer les détails d'une invitation selon un **id d'invitation** et un **code**\n",
         ],
         tags: ['api', 'invitations'],
-      },
-    },
-    {
-      method: 'DELETE',
-      path: '/api/certification-center-invitations/{certificationCenterInvitationId}',
-      config: {
-        handler: certificationCenterInvitationAdminController.cancelCertificationCenterInvitation,
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsAdminOfCertificationCenterWithCertificationCenterInvitationId,
-            assign: 'isAdminOfCertificationCenter',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            certificationCenterInvitationId: identifiersType.certificationCenterInvitationId.required(),
-          }),
-        },
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs appartenant à un centre de certification**\n',
-          "- Cette route permet d'annuler une invitation actuellement en attente selon un **id d'invitation**",
-        ],
-        tags: ['api', 'certification-center-invitation'],
       },
     },
     {
