+++ api/tests/team/acceptance/application/certification-center-invitation/certification-center-invitation.route.test.js
@@ -6,6 +6,8 @@ import {
   knex,
 } from '../../../../test-helper.js';
 
+const CERTIFICATION_CENTER_INVITATIONS_TABLE_NAME = 'certification-center-invitations';
+
 describe('Acceptance | Team | Application | Route | Certification Center Invitation', function () {
   let server, request;
 
@@ -78,6 +80,50 @@ describe('Acceptance | Team | Application | Route | Certification Center Invitat
           .whereIn('email', emails);
         expect(response.statusCode).to.equal(204);
         expect(certificationCenterInvitations.length).to.equal(2);
+      });
+    });
+  });
+
+  describe('DELETE /api/certification-center-invitations/{id}', function () {
+    context('when user is an admin of the certification center', function () {
+      let adminUser, certificationCenter;
+
+      beforeEach(async function () {
+        adminUser = databaseBuilder.factory.buildUser();
+        certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+        databaseBuilder.factory.buildCertificationCenterMembership({
+          userId: adminUser.id,
+          certificationCenterId: certificationCenter.id,
+          role: 'ADMIN',
+        });
+
+        await databaseBuilder.commit();
+      });
+
+      it('cancels the certification center invitation and returns a 204 HTTP status code', async function () {
+        // given
+        const certificationCenterInvitation = databaseBuilder.factory.buildCertificationCenterInvitation({
+          certificationCenterId: certificationCenter.id,
+        });
+        const request = {
+          headers: {
+            authorization: generateValidRequestAuthorizationHeader(adminUser.id),
+          },
+          method: 'DELETE',
+          url: `/api/certification-center-invitations/${certificationCenterInvitation.id}`,
+        };
+
+        await databaseBuilder.commit();
+
+        // when
+        const { statusCode } = await server.inject(request);
+
+        // then
+        const cancelledCertificationCenterInvitation = await knex(CERTIFICATION_CENTER_INVITATIONS_TABLE_NAME)
+          .where({ id: certificationCenterInvitation.id })
+          .first();
+        expect(cancelledCertificationCenterInvitation.status).to.equal('cancelled');
+        expect(statusCode).to.equal(204);
       });
     });
   });
